Add pipeline module
