+++ chain-signatures/node/src/stream/mod.rs
@@ -1,3 +1,6 @@
+pub mod ops;
+pub mod pipeline;
+
 use crate::backlog::Backlog;
 use crate::mesh::MeshState;
 use crate::node_client::NodeClient;
@@ -10,8 +13,7 @@ use crate::stream::ops::{
     process_sign_request, requeue_pending_sign_requests, resume_pending_publish_requests,
     RespondBidirectionalEvent, SignatureRespondedEvent,
 };
-
-pub mod ops;
+use crate::stream::pipeline::ChainPipeline;
 
 use async_trait::async_trait;
 use std::time::Duration;
@@ -173,212 +175,6 @@ pub trait ChainStream: Send + 'static {
 
     async fn start(&mut self) -> anyhow::Result<Self::Indexer>;
     async fn next_event(&mut self) -> Option<ChainEvent>;
-}
-
-pub struct ChainPipeline<I: ChainIndexer> {
-    indexer: I,
-    state_tx: watch::Sender<ChainStreaming>,
-    state_rx: watch::Receiver<ChainStreaming>,
-    backlog: Backlog,
-    checkpoints_rx: watch::Receiver<CheckpointDigest>,
-    mesh_state: watch::Receiver<MeshState>,
-    node_client: NodeClient,
-    threshold: usize,
-}
-
-impl<I: ChainIndexer> ChainPipeline<I> {
-    pub fn new(
-        indexer: I,
-        state_tx: watch::Sender<ChainStreaming>,
-        state_rx: watch::Receiver<ChainStreaming>,
-        backlog: Backlog,
-        checkpoints_rx: watch::Receiver<CheckpointDigest>,
-        mesh_state: watch::Receiver<MeshState>,
-        node_client: NodeClient,
-        threshold: usize,
-    ) -> Self {
-        Self {
-            indexer,
-            state_tx,
-            state_rx,
-            backlog,
-            checkpoints_rx,
-            mesh_state,
-            node_client,
-            threshold,
-        }
-    }
-
-    pub async fn run(mut self) {
-        let chain = I::CHAIN;
-        tracing::info!(%chain, "starting ChainStream pipeline");
-        let mut current_state = *self.state_rx.borrow_and_update();
-
-        loop {
-            match current_state {
-                ChainStreaming::Recovery => {
-                    if let Some(next_state) = self.handle_recovery().await {
-                        current_state = next_state;
-                    } else {
-                        break;
-                    }
-                }
-                ChainStreaming::Catchup { anchor_height } => {
-                    if let Some(next_state) = self.handle_catchup(anchor_height).await {
-                        current_state = next_state;
-                    } else {
-                        break;
-                    }
-                }
-                ChainStreaming::Live => {
-                    if let Some(next_state) = self.handle_live().await {
-                        current_state = next_state;
-                    } else {
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    async fn handle_recovery(&mut self) -> Option<ChainStreaming> {
-        let chain = I::CHAIN;
-        tracing::info!(%chain, "starting Recovery state");
-        crate::mesh::wait_threshold_active(&mut self.mesh_state.clone(), self.threshold).await;
-
-        // Load local checkpoint from storage first
-        match self.backlog.storage.load_latest(chain).await {
-            Ok(Some(checkpoint)) => {
-                tracing::info!(
-                    ?chain,
-                    height = checkpoint.height,
-                    "loaded local checkpoint"
-                );
-                if let Err(err) = self.backlog.recover_by_checkpoint(checkpoint).await {
-                    tracing::warn!(?chain, %err, "failed to recover from local checkpoint");
-                }
-            }
-            Ok(None) => {
-                tracing::info!(?chain, "no local checkpoint found");
-            }
-            Err(err) => {
-                tracing::warn!(?chain, %err, "failed to load local checkpoint");
-            }
-        }
-
-        // Perform consensus checkpoint alignment
-        let _ = crate::stream::ops::align_backlog_with_consensus(
-            chain,
-            &self.backlog,
-            &mut self.checkpoints_rx,
-            &self.mesh_state,
-            &self.node_client,
-        )
-        .await;
-
-        // Determine anchor height
-        let anchor_height = loop {
-            match self.indexer.livestream().await {
-                Ok(anchor_height) => break anchor_height,
-                Err(err) => {
-                    tracing::error!(?err, %chain, "failed to initialize livestream; retrying");
-                    tokio::time::sleep(I::RETRY_DELAY).await;
-                }
-            }
-        };
-
-        let catchup_height = anchor_height.unwrap_or(0);
-        let next_state = ChainStreaming::Catchup {
-            anchor_height: catchup_height,
-        };
-        let _ = self.state_tx.send(next_state);
-        Some(next_state)
-    }
-
-    async fn handle_catchup(&mut self, anchor_height: u64) -> Option<ChainStreaming> {
-        let chain = I::CHAIN;
-        tracing::info!(%chain, anchor_height, "starting/re-starting catchup");
-        let mut catchup_iter = self.indexer.catchup_range(anchor_height).await;
-        let mut catchup_done = false;
-
-        let current_state = ChainStreaming::Catchup { anchor_height };
-
-        while !catchup_done {
-            tokio::select! {
-                catchup_item = catchup_iter.next() => {
-                    let Some(catchup_item) = catchup_item else {
-                        catchup_done = true;
-                        break;
-                    };
-                    while let Err(err) = self.indexer.process_catchup(&catchup_item).await {
-                        tracing::warn!(?err, %chain, "catchup item processing failed; retrying");
-                        tokio::time::sleep(I::RETRY_DELAY).await;
-                    }
-                }
-                _ = self.state_rx.changed() => {
-                    let new_state = *self.state_rx.borrow_and_update();
-                    if new_state != current_state {
-                        return Some(new_state);
-                    }
-                }
-                _ = self.checkpoints_rx.changed() => {
-                    if let Some(new_state) = check_regression_and_get_state(
-                        chain,
-                        &self.backlog,
-                        &mut self.checkpoints_rx,
-                    ).await {
-                        let _ = self.state_tx.send(new_state);
-                        return Some(new_state);
-                    }
-                }
-            }
-        }
-
-        let mut final_state = current_state;
-        if catchup_done {
-            tracing::info!(%chain, "catchup completed => transitioning to livestream");
-            if let Err(err) = self.indexer.notify_catchup_completed().await {
-                tracing::warn!(?err, %chain, "failed to signal catchup completion");
-            }
-            final_state = ChainStreaming::Live;
-            let _ = self.state_tx.send(final_state);
-        }
-        Some(final_state)
-    }
-
-    async fn handle_live(&mut self) -> Option<ChainStreaming> {
-        let chain = I::CHAIN;
-        let current_state = ChainStreaming::Live;
-        loop {
-            tokio::select! {
-                block = self.indexer.next() => {
-                    let Some(block) = block else {
-                        return None; // shutdown
-                    };
-                    while let Err(err) = self.indexer.process(&block).await {
-                        tracing::warn!(?err, "live block processing failed; retrying");
-                        tokio::time::sleep(I::RETRY_DELAY).await;
-                    }
-                }
-                _ = self.state_rx.changed() => {
-                    let new_state = *self.state_rx.borrow_and_update();
-                    if new_state != current_state {
-                        return Some(new_state);
-                    }
-                }
-                _ = self.checkpoints_rx.changed() => {
-                    if let Some(new_state) = check_regression_and_get_state(
-                        chain,
-                        &self.backlog,
-                        &mut self.checkpoints_rx,
-                    ).await {
-                        let _ = self.state_tx.send(new_state);
-                        return Some(new_state);
-                    }
-                }
-            }
-        }
-    }
 }
 
 async fn check_regression_and_get_state(
